+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixes
 - Inject `FlexibleSearch` into Kotlin String template [#535](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/535)
 - Readonly mode is broken in IDEA 2023.2 [#533](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/533)
+- Custom addon module support [#538](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/538)
 
 ## [2023.2.5]